pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "amifuse"
dynamic = ["version"]
description = "Mount Amiga filesystem images via FUSE using native AmigaOS handlers"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "BSD-2-Clause"}
authors = [
    {name = "Stefan Reinauer"},
]
keywords = ["amiga", "fuse", "filesystem", "pfs3", "m68k"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: System :: Filesystems",
]

dependencies = [
    "fusepy>=3.0",
    "amitools-amifuse[vamos]>=0.8.0.post2",
]

[project.scripts]
amifuse = "amifuse.fuse_fs:main"
rdb-inspect = "amifuse.rdb_inspect:main"
driver-info = "amifuse.driver_info:main"

[project.urls]
Homepage = "https://github.com/reinauer/amifuse"

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-timeout>=2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "fuse: requires FUSE/WinFSP kernel driver",
    "integration: integration tests requiring real fixtures and machine68k",
    "smoke: smoke test wrappers for tools/ scripts (requires external fixtures)",
    "windows: Windows-specific tests",
    "macos: macOS-specific tests",
    "linux: Linux-specific tests",
]
filterwarnings = ["ignore::DeprecationWarning:amitools"]

[tool.setuptools.packages.find]
include = ["amifuse*"]

[tool.setuptools_scm]
local_scheme = "no-local-version"
